scripts/requirements.txt
distro==1.9.0
psutil==7.0.0
py-cpuinfo==9.0.0
